Widen PiResult.pi to u128 so pi(10^21) fits
pi(10^21) = 21,127,269,486,018,731,928 ~ 2.1e19 EXCEEDS u64 (max 1.84e19), so the
pi field must be u128 to reach 10^21 -- the value itself, not just x, crosses the
64-bit wall one power of ten after x does (x crosses at 10^20). The i128 accumulator
r = phi + a - 1 - p2 already had the range; only the result field was too narrow.

+++ src/lmo.zig
@@ -1427,7 +1427,7 @@ pub fn s2P2Parallel(comptime C: type, comptime X: type, gpa: std.mem.Allocator, 
 
 // ---------------------------------------------------------------------- π(x)
 
-pub const PiResult = struct { pi: u64, phi: i128, p2: i128, y: u64, a: usize, z: u64, leaves: u64 };
+pub const PiResult = struct { pi: u128, phi: i128, p2: i128, y: u64, a: usize, z: u64, leaves: u64 };
 
 /// π(x) = φ(x, a) + a − 1 − P₂(x, a), a = π(y) — LMO end to end.
 /// π(x) end to end. Takes u128 so x can exceed 2⁶⁴ (π(10²⁰)); dispatches to the u64
